consolidate 404 and unauthorized pages
