+++ frontend/src/metabase/App.jsx
@@ -7,8 +7,7 @@ import Navbar from "metabase/nav/containers/Navbar.jsx";
 
 import UndoListing from "metabase/containers/UndoListing";
 
-import NotFound from "metabase/components/NotFound.jsx";
-import Unauthorized from "metabase/components/Unauthorized.jsx";
+import { NotFound, Unauthorized } from "metabase/containers/ErrorPages";
 import Archived from "metabase/components/Archived.jsx";
 import GenericError from "metabase/components/GenericError.jsx";
